Complete Lab1

+++ lab1/app/src/main/java/com/example/lab1/Currency.kt
@@ -1,7 +1,7 @@
 package com.example.lab1
 
 data class Currency (
-    val USD: Float = 2.51f,
-    val EUR: Float = 2.48f,
-    val BYN: Float = 1.0f
+    val USD: Double = 2.51,
+    val EUR: Double = 2.48,
+    val BYN: Double = 1.0
 )
