+++ lab1/app/src/main/java/com/example/lab1/Distance.kt
@@ -1,7 +1,7 @@
 package com.example.lab1
 
 data class Distance (
-    val kilometers: Float = 1f,
-    val meters: Float = 1000f,
-    val centimeters: Float = 10000f,
+    val kilometers: Double = 1.0,
+    val meters: Double = 1000.0,
+    val centimeters: Double = 10000.0,
 )
